move test_rop_cache to integration test suite
