+++ test/integration/meson.build
@@ -27,6 +27,7 @@ if get_option('enable_tests') and cli_enabled
     'open_analyse_save_load_project',
     'pdb',
     'project_migrate',
+    'rop_cache',
     'rzpipe',
     'str_search',
   ]
