+++ test/unit/test_rop.c
@@ -141,70 +141,7 @@ bool test_rz_direct_solver() {
 	mu_end;
 }
 
-bool test_rop_cache(void) {
-	RzCore *core = rz_core_new();
-	mu_assert_notnull(core, "core");
-
-#if __WINDOWS__
-	const char *test_bin_name = "bins/pe/standard.exe";
-#else
-	const char *test_bin_name = "bins/elf/analysis/hello-linux-x86_64";
-#endif
-	char *test_bin = NULL;
-	if (rz_file_exists(test_bin_name)) {
-		test_bin = strdup(test_bin_name);
-	} else {
-		// Fallback for environments where bins/ is missing but rizin-testbins exists
-		// Try: rizin-testbins/elf/... (CWD is root)
-		test_bin = rz_str_newf("rizin-testbins/%s", test_bin_name + 5);
-		if (!rz_file_exists(test_bin)) {
-			free(test_bin);
-			// Try: ../rizin-testbins/elf/... (CWD is test/)
-			test_bin = rz_str_newf("../rizin-testbins/%s", test_bin_name + 5);
-			if (!rz_file_exists(test_bin)) {
-				free(test_bin);
-				// Try: ../../../rizin-testbins/elf/... (CWD is build/test/unit)
-				test_bin = rz_str_newf("../../../rizin-testbins/%s", test_bin_name + 5);
-			}
-		}
-	}
-
-	RzCoreFile *cf = rz_core_file_open(core, test_bin, RZ_PERM_RX, 0);
-	mu_assert_notnull(cf, "open file");
-	rz_core_bin_load(core, NULL, 0);
-	rz_config_set_b(core->config, "rop.cache", true);
-
-	const char *filter = "ret";
-	ut64 key = rz_str_djb2_hash(filter);
-
-	RzCmdStateOutput state;
-	rz_cmd_state_output_init(&state, RZ_OUTPUT_MODE_QUIET, core);
-
-	RzRopSearchContext *ctx = rz_core_rop_search_context_new(
-		core, filter, false, RZ_ROP_GADGET_PRINT, RZ_ROP_DETAIL_SEARCH_NON, &state);
-	rz_core_rop_search(core, ctx);
-	rz_core_rop_search_context_free(ctx);
-
-	mu_assert_notnull(core->analysis->ht_rop, "ht_rop");
-	char *result = ht_up_find(core->analysis->ht_rop, key, NULL);
-	mu_assert_notnull(result, "cached result");
-
-	ctx = rz_core_rop_search_context_new(
-		core, filter, false, RZ_ROP_GADGET_PRINT, RZ_ROP_DETAIL_SEARCH_NON, &state);
-	rz_core_rop_search(core, ctx);
-	rz_core_rop_search_context_free(ctx);
-
-	char *result2 = ht_up_find(core->analysis->ht_rop, key, NULL);
-	mu_assert_ptreq(result2, result, "cache hit");
-
-	free(test_bin);
-	rz_cmd_state_output_fini(&state);
-	rz_core_free(core);
-	mu_end;
-}
-
 bool all_tests() {
-	mu_run_test(test_rop_cache);
 	mu_run_test(test_rz_direct_solver);
 	return tests_passed != tests_run;
 }
